requirements.txt
# Agent required
bitarray==1.2.1
nltk==3.4.5
ruamel.yaml==0.16.10
fire==0.2.1
tqdm==4.43.0
watchdog==0.10.2
scipy==1.4.1
numpy==1.18.1

# textworld required
gym==0.17.0
TatSu==4.4.0
textworld==1.2.0

# NN required
tensorflow==1.14.1
bert-tensorflow==1.0.1
albert-tensorflow==1.1
sentencepiece==0.1.85
